requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8.4
django-annoying==0.8.7
django-toolbelt==0.0.1
gunicorn==18.0
jsonfield==1.0.3
psycopg2==2.5.3
pystache==0.5.4
static==1.0.2
static3==0.6.1
